+++ aom_util/debug_util.h
@@ -49,16 +49,16 @@ void bitstream_queue_set_skip_read(int skip);
 void mismatch_move_frame_idx_w();
 void mismatch_move_frame_idx_r();
 void mismatch_reset_frame(int num_planes);
-void mismatch_record_block_pre(const uint8_t *src, int src_stride,
+void mismatch_record_block_pre(const uint16_t *src, int src_stride,
                                int frame_offset, int plane, int pixel_c,
                                int pixel_r, int blk_w, int blk_h);
-void mismatch_record_block_tx(const uint8_t *src, int src_stride,
+void mismatch_record_block_tx(const uint16_t *src, int src_stride,
                               int frame_offset, int plane, int pixel_c,
                               int pixel_r, int blk_w, int blk_h);
-void mismatch_check_block_pre(const uint8_t *src, int src_stride,
+void mismatch_check_block_pre(const uint16_t *src, int src_stride,
                               int frame_offset, int plane, int pixel_c,
                               int pixel_r, int blk_w, int blk_h);
-void mismatch_check_block_tx(const uint8_t *src, int src_stride,
+void mismatch_check_block_tx(const uint16_t *src, int src_stride,
                              int frame_offset, int plane, int pixel_c,
                              int pixel_r, int blk_w, int blk_h);
 #endif  // CONFIG_MISMATCH_DEBUG
